+++ HaplotypePhaserSym.h
@@ -57,7 +57,7 @@ public:
 	Pedigree ped;
 	float theta;
 	float error;
-	int Ne;
+	float Ne;
 
 	std::vector<double> distances;
